[mypyc] Refactor: extract class and module IR from mypyc.ops
These aren't really "ops".
