+++ mypyc/emit.py
@@ -12,9 +12,10 @@ from mypyc.ops import (
     RUnion, RPrimitive,
     is_float_rprimitive, is_bool_rprimitive, is_int_rprimitive, is_short_int_rprimitive,
     is_list_rprimitive, is_dict_rprimitive, is_set_rprimitive, is_tuple_rprimitive,
-    is_none_rprimitive, is_object_rprimitive, object_rprimitive, is_str_rprimitive, ClassIR,
-    FuncDecl, int_rprimitive, is_optional_type, optional_value_type, all_concrete_classes
+    is_none_rprimitive, is_object_rprimitive, object_rprimitive, is_str_rprimitive,
+    FuncDecl, int_rprimitive, is_optional_type, optional_value_type
 )
+from mypyc.class_ir import ClassIR, all_concrete_classes
 from mypyc.namegen import NameGenerator, exported_name
 from mypyc.sametype import is_same_type
