+++ mypyc/irbuild/mapper.py
@@ -11,8 +11,9 @@ from mypy.types import (
 from mypyc.ops import (
     RType, RUnion, RTuple, RInstance, object_rprimitive, dict_rprimitive, tuple_rprimitive,
     none_rprimitive, int_rprimitive, float_rprimitive, str_rprimitive, bool_rprimitive,
-    list_rprimitive, set_rprimitive, FuncSignature, ClassIR, FuncDecl, RuntimeArg, LiteralsMap
+    list_rprimitive, set_rprimitive, FuncSignature, FuncDecl, RuntimeArg, LiteralsMap
 )
+from mypyc.class_ir import ClassIR
 
 
 class Mapper:
