+++ mypyc/ops.py
@@ -15,15 +15,17 @@ from typing import (
     List, Sequence, Dict, Generic, TypeVar, Optional, Any, NamedTuple, Tuple, Callable,
     Union, Iterable, Set
 )
-from typing_extensions import Final, Type, ClassVar
 from collections import OrderedDict
 
-from mypy.nodes import ARG_NAMED_OPT, ARG_OPT, ARG_POS, Block, FuncDef, SymbolNode
-from mypyc.common import PROPSET_PREFIX
-
+from typing_extensions import Final, Type, ClassVar, TYPE_CHECKING
 from mypy_extensions import trait
 
-from mypyc.namegen import NameGenerator, exported_name
+from mypy.nodes import ARG_NAMED_OPT, ARG_OPT, ARG_POS, Block, FuncDef, SymbolNode
+
+from mypyc.namegen import NameGenerator
+
+if TYPE_CHECKING:
+    from mypyc.class_ir import ClassIR
 
 T = TypeVar('T')
 
@@ -1636,483 +1638,7 @@ class FuncIR:
 INVALID_FUNC_DEF = FuncDef('<INVALID_FUNC_DEF>', [], Block([]))  # type: Final
 
 
-# Some notes on the vtable layout: Each concrete class has a vtable
-# that contains function pointers for its methods. So that subclasses
-# may be efficiently used when their parent class is expected, the
-# layout of child vtables must be an extension of their base class's
-# vtable.
-#
-# This makes multiple inheritance tricky, since obviously we cannot be
-# an extension of multiple parent classes. We solve this by requriing
-# all but one parent to be "traits", which we can operate on in a
-# somewhat less efficient way. For each trait implemented by a class,
-# we generate a separate vtable for the methods in that trait.
-# We then store an array of (trait type, trait vtable) pointers alongside
-# a class's main vtable. When we want to call a trait method, we
-# (at runtime!) search the array of trait vtables to find the correct one,
-# then call through it.
-# Trait vtables additionally need entries for attribute getters and setters,
-# since they can't always be in the same location.
-#
-# To keep down the number of indirections necessary, we store the
-# array of trait vtables in the memory *before* the class vtable, and
-# search it backwards.  (This is a trick we can only do once---there
-# are only two directions to store data in---but I don't think we'll
-# need it again.)
-# There are some tricks we could try in the future to store the trait
-# vtables inline in the trait table (which would cut down one indirection),
-# but this seems good enough for now.
-#
-# As an example:
-# Imagine that we have a class B that inherits from a concrete class A
-# and traits T1 and T2, and that A has methods foo() and
-# bar() and B overrides bar() with a more specific type.
-# Then B's vtable will look something like:
-#
-#      T1 type object
-#      ptr to B's T1 trait vtable
-#      T2 type object
-#      ptr to B's T2 trait vtable
-# -> | A.foo
-#    | Glue function that converts between A.bar's type and B.bar
-#      B.bar
-#      B.baz
-#
-# The arrow points to the "start" of the vtable (what vtable pointers
-# point to) and the bars indicate which parts correspond to the parent
-# class A's vtable layout.
-#
-# Classes that allow interpreted code to subclass them also have a
-# "shadow vtable" that contains implementations that delegate to
-# making a pycall, so that overridden methods in interpreted children
-# will be called. (A better strategy could dynamically generate these
-# vtables based on which methods are overridden in the children.)
-
-# Descriptions of method and attribute entries in class vtables.
-# The 'cls' field is the class that the method/attr was defined in,
-# which might be a parent class.
-# The 'shadow_method', if present, contains the method that should be
-# placed in the class's shadow vtable (if it has one).
-
-VTableMethod = NamedTuple(
-    'VTableMethod', [('cls', 'ClassIR'),
-                     ('name', str),
-                     ('method', FuncIR),
-                     ('shadow_method', Optional[FuncIR])])
-
-
-VTableAttr = NamedTuple(
-    'VTableAttr', [('cls', 'ClassIR'),
-                   ('name', str),
-                   ('is_setter', bool)])
-
-
-VTableEntry = Union[VTableMethod, VTableAttr]
-VTableEntries = List[VTableEntry]
-
-
-def serialize_vtable_entry(entry: VTableEntry) -> JsonDict:
-    if isinstance(entry, VTableMethod):
-        return {
-            '.class': 'VTableMethod',
-            'cls': entry.cls.fullname,
-            'name': entry.name,
-            'method': entry.method.decl.fullname,
-            'shadow_method': entry.shadow_method.decl.fullname if entry.shadow_method else None,
-        }
-    else:
-        return {
-            '.class': 'VTableAttr',
-            'cls': entry.cls.fullname,
-            'name': entry.name,
-            'is_setter': entry.is_setter,
-        }
-
-
-def serialize_vtable(vtable: VTableEntries) -> List[JsonDict]:
-    return [serialize_vtable_entry(v) for v in vtable]
-
-
-def deserialize_vtable_entry(data: JsonDict, ctx: DeserMaps) -> VTableEntry:
-    if data['.class'] == 'VTableMethod':
-        return VTableMethod(
-            ctx.classes[data['cls']], data['name'], ctx.functions[data['method']],
-            ctx.functions[data['shadow_method']] if data['shadow_method'] else None)
-    elif data['.class'] == 'VTableAttr':
-        return VTableAttr(ctx.classes[data['cls']], data['name'], data['is_setter'])
-    assert False, "Bogus vtable .class: %s" % data['.class']
-
-
-def deserialize_vtable(data: List[JsonDict], ctx: DeserMaps) -> VTableEntries:
-    return [deserialize_vtable_entry(x, ctx) for x in data]
-
-
-class ClassIR:
-    """Intermediate representation of a class.
-
-    This also describes the runtime structure of native instances.
-    """
-    def __init__(self, name: str, module_name: str, is_trait: bool = False,
-                 is_generated: bool = False, is_abstract: bool = False,
-                 is_ext_class: bool = True) -> None:
-        self.name = name
-        self.module_name = module_name
-        self.is_trait = is_trait
-        self.is_generated = is_generated
-        self.is_abstract = is_abstract
-        self.is_ext_class = is_ext_class
-        # An augmented class has additional methods separate from what mypyc generates.
-        # Right now the only one is dataclasses.
-        self.is_augmented = False
-        self.inherits_python = False
-        self.has_dict = False
-        # Do we allow interpreted subclasses? Derived from a mypyc_attr.
-        self.allow_interpreted_subclasses = False
-        # If this a subclass of some built-in python class, the name
-        # of the object for that class. We currently only support this
-        # in a few ad-hoc cases.
-        self.builtin_base = None  # type: Optional[str]
-        # Default empty ctor
-        self.ctor = FuncDecl(name, None, module_name, FuncSignature([], RInstance(self)))
-
-        self.attributes = OrderedDict()  # type: OrderedDict[str, RType]
-        # We populate method_types with the signatures of every method before
-        # we generate methods, and we rely on this information being present.
-        self.method_decls = OrderedDict()  # type: OrderedDict[str, FuncDecl]
-        # Map of methods that are actually present in an extension class
-        self.methods = OrderedDict()  # type: OrderedDict[str, FuncIR]
-        # Glue methods for boxing/unboxing when a class changes the type
-        # while overriding a method. Maps from (parent class overrided, method)
-        # to IR of glue method.
-        self.glue_methods = OrderedDict()  # type: Dict[Tuple[ClassIR, str], FuncIR]
-
-        # Properties are accessed like attributes, but have behavior like method calls.
-        # They don't belong in the methods dictionary, since we don't want to expose them to
-        # Python's method API. But we want to put them into our own vtable as methods, so that
-        # they are properly handled and overridden. The property dictionary values are a tuple
-        # containing a property getter and an optional property setter.
-        self.properties = OrderedDict()  # type: OrderedDict[str, Tuple[FuncIR, Optional[FuncIR]]]
-        # We generate these in prepare_class_def so that we have access to them when generating
-        # other methods and properties that rely on these types.
-        self.property_types = OrderedDict()  # type: OrderedDict[str, RType]
-
-        self.vtable = None  # type: Optional[Dict[str, int]]
-        self.vtable_entries = []  # type: VTableEntries
-        self.trait_vtables = OrderedDict()  # type: OrderedDict[ClassIR, VTableEntries]
-        # N.B: base might not actually quite be the direct base.
-        # It is the nearest concrete base, but we allow a trait in between.
-        self.base = None  # type: Optional[ClassIR]
-        self.traits = []  # type: List[ClassIR]
-        # Supply a working mro for most generated classes. Real classes will need to
-        # fix it up.
-        self.mro = [self]  # type: List[ClassIR]
-        # base_mro is the chain of concrete (non-trait) ancestors
-        self.base_mro = [self]  # type: List[ClassIR]
-
-        # Direct subclasses of this class (use subclasses() to also incude non-direct ones)
-        # None if separate compilation prevents this from working
-        self.children = []  # type: Optional[List[ClassIR]]
-
-    @property
-    def fullname(self) -> str:
-        return "{}.{}".format(self.module_name, self.name)
-
-    def real_base(self) -> Optional['ClassIR']:
-        """Return the actual concrete base class, if there is one."""
-        if len(self.mro) > 1 and not self.mro[1].is_trait:
-            return self.mro[1]
-        return None
-
-    def vtable_entry(self, name: str) -> int:
-        assert self.vtable is not None, "vtable not computed yet"
-        assert name in self.vtable, '%r has no attribute %r' % (self.name, name)
-        return self.vtable[name]
-
-    def attr_details(self, name: str) -> Tuple[RType, 'ClassIR']:
-        for ir in self.mro:
-            if name in ir.attributes:
-                return ir.attributes[name], ir
-            if name in ir.property_types:
-                return ir.property_types[name], ir
-        raise KeyError('%r has no attribute %r' % (self.name, name))
-
-    def attr_type(self, name: str) -> RType:
-        return self.attr_details(name)[0]
-
-    def method_decl(self, name: str) -> FuncDecl:
-        for ir in self.mro:
-            if name in ir.method_decls:
-                return ir.method_decls[name]
-        raise KeyError('%r has no attribute %r' % (self.name, name))
-
-    def method_sig(self, name: str) -> FuncSignature:
-        return self.method_decl(name).sig
-
-    def has_method(self, name: str) -> bool:
-        try:
-            self.method_decl(name)
-        except KeyError:
-            return False
-        return True
-
-    def is_method_final(self, name: str) -> bool:
-        subs = self.subclasses()
-        if subs is None:
-            # TODO: Look at the final attribute!
-            return False
-
-        if self.has_method(name):
-            method_decl = self.method_decl(name)
-            for subc in subs:
-                if subc.method_decl(name) != method_decl:
-                    return False
-            return True
-        else:
-            return not any(subc.has_method(name) for subc in subs)
-
-    def has_attr(self, name: str) -> bool:
-        try:
-            self.attr_type(name)
-        except KeyError:
-            return False
-        return True
-
-    def name_prefix(self, names: NameGenerator) -> str:
-        return names.private_name(self.module_name, self.name)
-
-    def struct_name(self, names: NameGenerator) -> str:
-        return '{}Object'.format(exported_name(self.fullname))
-
-    def get_method_and_class(self, name: str) -> Optional[Tuple[FuncIR, 'ClassIR']]:
-        for ir in self.mro:
-            if name in ir.methods:
-                return ir.methods[name], ir
-
-        return None
-
-    def get_method(self, name: str) -> Optional[FuncIR]:
-        res = self.get_method_and_class(name)
-        return res[0] if res else None
-
-    def subclasses(self) -> Optional[Set['ClassIR']]:
-        """Return all subclassses of this class, both direct and indirect.
-
-        Return None if it is impossible to identify all subclasses, for example
-        because we are performing separate compilation.
-        """
-        if self.children is None or self.allow_interpreted_subclasses:
-            return None
-        result = set(self.children)
-        for child in self.children:
-            if child.children:
-                child_subs = child.subclasses()
-                if child_subs is None:
-                    return None
-                result.update(child_subs)
-        return result
-
-    def concrete_subclasses(self) -> Optional[List['ClassIR']]:
-        """Return all concrete (i.e. non-trait and non-abstract) subclasses.
-
-        Include both direct and indirect subclasses. Place classes with no children first.
-        """
-        subs = self.subclasses()
-        if subs is None:
-            return None
-        concrete = {c for c in subs if not (c.is_trait or c.is_abstract)}
-        # We place classes with no children first because they are more likely
-        # to appear in various isinstance() checks. We then sort leafs by name
-        # to get stable order.
-        return sorted(concrete, key=lambda c: (len(c.children or []), c.name))
-
-    def serialize(self) -> JsonDict:
-        return {
-            'name': self.name,
-            'module_name': self.module_name,
-            'is_trait': self.is_trait,
-            'is_ext_class': self.is_ext_class,
-            'is_abstract': self.is_abstract,
-            'is_generated': self.is_generated,
-            'is_augmented': self.is_augmented,
-            'inherits_python': self.inherits_python,
-            'has_dict': self.has_dict,
-            'allow_interpreted_subclasses': self.allow_interpreted_subclasses,
-            'builtin_base': self.builtin_base,
-            'ctor': self.ctor.serialize(),
-            # We serialize dicts as lists to ensure order is preserved
-            'attributes': [(k, t.serialize()) for k, t in self.attributes.items()],
-            # We try to serialize a name reference, but if the decl isn't in methods
-            # then we can't be sure that will work so we serialize the whole decl.
-            'method_decls': [(k, d.fullname if k in self.methods else d.serialize())
-                             for k, d in self.method_decls.items()],
-            # We serialize method fullnames out and put methods in a separate dict
-            'methods': [(k, m.fullname) for k, m in self.methods.items()],
-            'glue_methods': [
-                ((cir.fullname, k), m.fullname)
-                for (cir, k), m in self.glue_methods.items()
-            ],
-
-            # We serialize properties and property_types separately out of an
-            # abundance of caution about preserving dict ordering...
-            'property_types': [(k, t.serialize()) for k, t in self.property_types.items()],
-            'properties': list(self.properties),
-
-            'vtable': self.vtable,
-            'vtable_entries': serialize_vtable(self.vtable_entries),
-            'trait_vtables': [
-                (cir.fullname, serialize_vtable(v)) for cir, v in self.trait_vtables.items()
-            ],
-
-            # References to class IRs are all just names
-            'base': self.base.fullname if self.base else None,
-            'traits': [cir.fullname for cir in self.traits],
-            'mro': [cir.fullname for cir in self.mro],
-            'base_mro': [cir.fullname for cir in self.base_mro],
-            'children': [
-                cir.fullname for cir in self.children
-            ] if self.children is not None else None,
-        }
-
-    @classmethod
-    def deserialize(cls, data: JsonDict, ctx: DeserMaps) -> 'ClassIR':
-        fullname = data['module_name'] + '.' + data['name']
-        assert fullname in ctx.classes, "Class %s not in deser class map" % fullname
-        ir = ctx.classes[fullname]
-
-        ir.is_trait = data['is_trait']
-        ir.is_generated = data['is_generated']
-        ir.is_abstract = data['is_abstract']
-        ir.is_ext_class = data['is_ext_class']
-        ir.is_augmented = data['is_augmented']
-        ir.inherits_python = data['inherits_python']
-        ir.has_dict = data['has_dict']
-        ir.allow_interpreted_subclasses = data['allow_interpreted_subclasses']
-        ir.builtin_base = data['builtin_base']
-        ir.ctor = FuncDecl.deserialize(data['ctor'], ctx)
-        ir.attributes = OrderedDict(
-            (k, deserialize_type(t, ctx)) for k, t in data['attributes']
-        )
-        ir.method_decls = OrderedDict((k, ctx.functions[v].decl
-                                       if isinstance(v, str) else FuncDecl.deserialize(v, ctx))
-                                      for k, v in data['method_decls'])
-        ir.methods = OrderedDict((k, ctx.functions[v]) for k, v in data['methods'])
-        ir.glue_methods = OrderedDict(
-            ((ctx.classes[c], k), ctx.functions[v]) for (c, k), v in data['glue_methods']
-        )
-        ir.property_types = OrderedDict(
-            (k, deserialize_type(t, ctx)) for k, t in data['property_types']
-        )
-        ir.properties = OrderedDict(
-            (k, (ir.methods[k], ir.methods.get(PROPSET_PREFIX + k))) for k in data['properties']
-        )
-
-        ir.vtable = data['vtable']
-        ir.vtable_entries = deserialize_vtable(data['vtable_entries'], ctx)
-        ir.trait_vtables = OrderedDict(
-            (ctx.classes[k], deserialize_vtable(v, ctx)) for k, v in data['trait_vtables']
-        )
-
-        base = data['base']
-        ir.base = ctx.classes[base] if base else None
-        ir.traits = [ctx.classes[s] for s in data['traits']]
-        ir.mro = [ctx.classes[s] for s in data['mro']]
-        ir.base_mro = [ctx.classes[s] for s in data['base_mro']]
-        ir.children = data['children'] and [ctx.classes[s] for s in data['children']]
-
-        return ir
-
-
-class NonExtClassInfo:
-    """Information needed to construct a non-extension class.
-
-
-    Includes the class dictionary, a tuple of base classes,
-    the class annotations dictionary, and the metaclass.
-    """
-
-    def __init__(self, dict: Value, bases: Value, anns: Value, metaclass: Value) -> None:
-        self.dict = dict
-        self.bases = bases
-        self.anns = anns
-        self.metaclass = metaclass
-
-
 LiteralsMap = Dict[Tuple[Type[object], Union[int, float, str, bytes, complex]], str]
-
-
-class ModuleIR:
-    """Intermediate representation of a module."""
-
-    def __init__(
-            self,
-            fullname: str,
-            imports: List[str],
-            functions: List[FuncIR],
-            classes: List[ClassIR],
-            final_names: List[Tuple[str, RType]]) -> None:
-        self.fullname = fullname
-        self.imports = imports[:]
-        self.functions = functions
-        self.classes = classes
-        self.final_names = final_names
-
-    def serialize(self) -> JsonDict:
-        return {
-            'fullname': self.fullname,
-            'imports': self.imports,
-            'functions': [f.serialize() for f in self.functions],
-            'classes': [c.serialize() for c in self.classes],
-            'final_names': [(k, t.serialize()) for k, t in self.final_names],
-        }
-
-    @classmethod
-    def deserialize(cls, data: JsonDict, ctx: DeserMaps) -> 'ModuleIR':
-        return ModuleIR(
-            data['fullname'],
-            data['imports'],
-            [ctx.functions[FuncDecl.get_name_from_json(f['decl'])] for f in data['functions']],
-            [ClassIR.deserialize(c, ctx) for c in data['classes']],
-            [(k, deserialize_type(t, ctx)) for k, t in data['final_names']],
-        )
-
-
-def deserialize_modules(data: Dict[str, JsonDict], ctx: DeserMaps) -> Dict[str, ModuleIR]:
-    """Deserialize a collection of modules.
-
-    The modules can contain dependencies on each other.
-
-    Arguments:
-        data: A dict containing the modules to deserialize.
-        ctx: The deserialization maps to use and to populate.
-             They are populated with information from the deserialized
-             modules and as a precondition must have been populated by
-             deserializing any dependencies of the modules being deserialized
-             (outside of dependencies between the modules themselves).
-
-    Returns a map containing the deserialized modules.
-    """
-    for mod in data.values():
-        # First create ClassIRs for every class so that we can construct types and whatnot
-        for cls in mod['classes']:
-            ir = ClassIR(cls['name'], cls['module_name'])
-            assert ir.fullname not in ctx.classes, "Class %s already in map" % ir.fullname
-            ctx.classes[ir.fullname] = ir
-
-    for mod in data.values():
-        # Then deserialize all of the functions so that methods are available
-        # to the class deserialization.
-        for method in mod['functions']:
-            func = FuncIR.deserialize(method, ctx)
-            assert func.decl.fullname not in ctx.functions, (
-                "Method %s already in map" % func.decl.fullname)
-            ctx.functions[func.decl.fullname] = func
-
-    return {k: ModuleIR.deserialize(v, ctx) for k, v in data.items()}
-
-
-# ModulesIRs should also always be an *OrderedDict*, but if we
-# declared it that way we would need to put it in quotes everywhere...
-ModuleIRs = Dict[str, ModuleIR]
 
 
 @trait
@@ -2249,25 +1775,6 @@ def format_func(fn: FuncIR) -> List[str]:
     code = format_blocks(fn.blocks, fn.env)
     lines.extend(code)
     return lines
-
-
-def format_modules(modules: ModuleIRs) -> List[str]:
-    ops = []
-    for module in modules.values():
-        for fn in module.functions:
-            ops.extend(format_func(fn))
-            ops.append('')
-    return ops
-
-
-def all_concrete_classes(class_ir: ClassIR) -> Optional[List[ClassIR]]:
-    """Return all concrete classes among the class itself and its subclasses."""
-    concrete = class_ir.concrete_subclasses()
-    if concrete is None:
-        return None
-    if not (class_ir.is_abstract or class_ir.is_trait):
-        concrete.append(class_ir)
-    return concrete
 
 
 def short_name(name: str) -> str:
